fix(argocd): address NFS migration review feedback
