--- workspaces/argocd/packages/app/package.json
@@ -1,86 +1,0 @@
-{
-  "name": "app",
-  "version": "0.0.9",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "license": "Apache-2.0",
-  "private": true,
-  "bundled": true,
-  "backstage": {
-    "role": "frontend"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/backstage/community-plugins",
-    "directory": "workspaces/argocd/packages/app"
-  },
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "tsc": "tsc",
-    "clean": "backstage-cli package clean",
-    "test": "backstage-cli package test --passWithNoTests --coverage",
-    "lint": "backstage-cli package lint"
-  },
-  "dependencies": {
-    "@backstage-community/plugin-argocd": "workspace:^",
-    "@backstage-community/plugin-rbac": "^1.47.1",
-    "@backstage/app-defaults": "^1.7.9",
-    "@backstage/catalog-model": "^1.9.0",
-    "@backstage/core-app-api": "^1.20.2",
-    "@backstage/core-components": "^0.18.11",
-    "@backstage/core-plugin-api": "^1.12.7",
-    "@backstage/integration-react": "^1.2.19",
-    "@backstage/plugin-api-docs": "^0.14.2",
-    "@backstage/plugin-catalog": "^2.0.6",
-    "@backstage/plugin-catalog-common": "^1.1.10",
-    "@backstage/plugin-catalog-graph": "^0.6.5",
-    "@backstage/plugin-catalog-import": "^0.13.14",
-    "@backstage/plugin-catalog-react": "^3.1.0",
-    "@backstage/plugin-kubernetes": "^0.12.20",
-    "@backstage/plugin-org": "^0.7.5",
-    "@backstage/plugin-permission-react": "^0.5.2",
-    "@backstage/plugin-scaffolder": "^1.38.0",
-    "@backstage/plugin-search": "^1.7.5",
-    "@backstage/plugin-search-react": "^1.11.5",
-    "@backstage/plugin-techdocs": "^1.17.7",
-    "@backstage/plugin-techdocs-module-addons-contrib": "^1.1.37",
-    "@backstage/plugin-techdocs-react": "^1.3.12",
-    "@backstage/plugin-user-settings": "^0.9.4",
-    "@backstage/ui": "^0.16.0",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.13.0",
-    "@mui/icons-material": "^5.15.16",
-    "@mui/material": "^5.15.16",
-    "@mui/styles": "^5.15.16",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-router": "^6.30.2",
-    "react-router-dom": "^6.23.0",
-    "tss-react": "^4.9.10"
-  },
-  "devDependencies": {
-    "@backstage/cli": "^0.36.3",
-    "@testing-library/dom": "^10.4.1",
-    "@testing-library/jest-dom": "6.9.1",
-    "@testing-library/react": "16.3.2",
-    "@types/node": "22.20.1",
-    "@types/react": "18.3.31",
-    "@types/react-dom": "18.3.7"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "files": [
-    "dist"
-  ]
-}
