--- workspaces/argocd/packages/backend/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "backend",
-  "version": "0.0.0",
-  "main": "dist/index.cjs.js",
-  "types": "src/index.ts",
-  "private": true,
-  "backstage": {
-    "role": "backend"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/backstage/community-plugins",
-    "directory": "workspaces/argocd/packages/backend"
-  },
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "tsc": "tsc",
-    "lint": "backstage-cli package lint",
-    "test": "backstage-cli package test --passWithNoTests --coverage",
-    "clean": "backstage-cli package clean"
-  },
-  "dependencies": {
-    "@backstage-community/plugin-argocd-backend": "workspace:^",
-    "@backstage-community/plugin-rbac-backend": "^7.6.1",
-    "@backstage/backend-defaults": "^0.17.3",
-    "@backstage/cli-node": "^0.3.3",
-    "@backstage/plugin-app-backend": "^0.5.15",
-    "@backstage/plugin-auth-backend": "^0.29.1",
-    "@backstage/plugin-auth-backend-module-github-provider": "^0.5.4",
-    "@backstage/plugin-auth-backend-module-guest-provider": "^0.2.20",
-    "@backstage/plugin-catalog-backend": "^3.8.0",
-    "@backstage/plugin-catalog-backend-module-logs": "^0.1.23",
-    "@backstage/plugin-catalog-backend-module-scaffolder-entity-model": "^0.2.21",
-    "@backstage/plugin-kubernetes-backend": "^0.21.5",
-    "@backstage/plugin-permission-backend": "^0.7.13",
-    "@backstage/plugin-proxy-backend": "^0.6.14",
-    "@backstage/plugin-scaffolder-backend": "^4.0.1",
-    "@backstage/plugin-search-backend": "^2.1.3",
-    "@backstage/plugin-search-backend-module-catalog": "^0.3.16",
-    "@backstage/plugin-search-backend-module-pg": "^0.5.56",
-    "@backstage/plugin-search-backend-module-techdocs": "^0.4.15",
-    "@backstage/plugin-techdocs-backend": "^2.2.1",
-    "app": "link:../app"
-  },
-  "devDependencies": {
-    "@backstage/cli": "^0.36.3"
-  },
-  "files": [
-    "dist"
-  ]
-}
