+++ workspaces/argocd/plugins/argocd/README.md
@@ -125,13 +125,30 @@ argocd:
 
 ## How to add Argo CD frontend plugin to Backstage app
 
-1. Install the Argo CD plugin using the following command:
+Install the Argo CD plugin using the following command:
 
 ```bash
 yarn workspace app add @backstage-community/plugin-argocd
 ```
 
-2. Add deployment summary and deployment lifecycle compoennt to the `entityPage.tsx` source file:
+### New frontend system
+
+If you are using Backstage's [new frontend system](https://backstage.io/docs/frontend-system/), add the default export to your app `features` array. The Deployment Lifecycle and Deployment Summary entity tabs are registered automatically for entities with Argo CD annotations.
+
+```ts
+import argocdPlugin from '@backstage-community/plugin-argocd';
+
+export const app = createApp({
+  features: [
+    // ...
+    argocdPlugin,
+  ],
+});
+```
+
+### Legacy frontend system
+
+If you are using the legacy frontend system, import the plugin from the `/legacy` subpath and add the deployment summary and deployment lifecycle components to the `entityPage.tsx` source file:
 
 ```ts
 // packages/app/src/components/catalog/EntityPage.tsx
@@ -139,7 +156,7 @@ import {
   ArgocdDeploymentSummary,
   ArgocdDeploymentLifecycle,
   isArgocdConfigured,
-} from '@backstage-community/plugin-argocd';
+} from '@backstage-community/plugin-argocd/legacy';
 
 const overviewContent = (
   <Grid container spacing={3} alignItems="stretch">
