+++ workspaces/argocd/plugins/argocd/dev/index.tsx
@@ -32,165 +32,33 @@ import { permissionApiRef } from '@backstage/plugin-permission-react';
 import userSettingsPlugin from '@backstage/plugin-user-settings/alpha';
 
 import {
-  KubernetesApi,
   kubernetesApiRef,
   kubernetesAuthProvidersApiRef,
 } from '@backstage/plugin-kubernetes-react';
 
 import { mockArgoResources } from './__data__/argoRolloutsObjects';
-import { customResourceTypes } from '../src/types/resources';
+import {
+  MockArgoCDApiClient,
+  MockKubernetesClient,
+  mockKubernetesAuthProviderApi,
+} from './__fixtures__/mockClients';
 
 import {
-  ArgoCDApi,
   argoCDApiRef,
-  ArgoCDAppDeployRevisionDetails,
   ArgoCDInstanceApiClient,
   argoCDInstanceApiRef,
-  FindApplicationsOptions,
-  GetApplicationOptions,
-  ListAppsOptions,
-  RevisionDetailsListOptions,
-  RevisionDetailsOptions,
 } from '../src/api';
 import argocdPlugin from '../src/plugin';
-import {
-  Application,
-  InstanceApplications,
-} from '@backstage-community/plugin-argocd-common';
 import {
   mockArgocdConfig,
   mockArgocdMultiInstanceConfig,
   mockEntity,
-  mockIdRevisions,
-  DEV_INSTANCE_APPLICATIONS,
   mockArgoMultiInstanceSelectorEntity,
   mockArgoMultiInstanceAppNameEntity,
   mockArgoOneAppEntity,
 } from './__data__';
 import { getArgocdInstances } from '../src/hooks/useArgocdConfig';
 import { ConfigReader } from '@backstage/config';
-
-const getInstanceNameFromUrl = (url: string): string => {
-  return url.replace('/argoInstance/', '');
-};
-
-class MockArgoCDApiClient implements ArgoCDApi {
-  async listApps(options: ListAppsOptions): Promise<{ items: Application[] }> {
-    const instanceName = getInstanceNameFromUrl(options.url);
-    let apps = DEV_INSTANCE_APPLICATIONS[instanceName] ?? [];
-
-    if (options.appSelector) {
-      const decodedSelector = decodeURIComponent(options.appSelector);
-      const [labelKey, labelValue] = decodedSelector.split('=', 2);
-      apps = apps.filter(app => app.metadata.labels?.[labelKey] === labelValue);
-    }
-
-    return { items: apps };
-  }
-
-  async getRevisionDetails(
-    options: RevisionDetailsOptions,
-  ): Promise<ArgoCDAppDeployRevisionDetails> {
-    return mockIdRevisions[options.revisionID];
-  }
-
-  async getRevisionDetailsList(
-    options: RevisionDetailsListOptions,
-  ): Promise<ArgoCDAppDeployRevisionDetails[]> {
-    if (!options.revisionIDs || options.revisionIDs.length < 1) {
-      return [];
-    }
-    const promises: Promise<ArgoCDAppDeployRevisionDetails>[] = [];
-
-    options.revisionIDs.forEach((revisionID: string) => {
-      const application = options.apps.find(app =>
-        app?.status?.history?.find(h => h.revision === revisionID),
-      );
-
-      if (application) {
-        promises.push(
-          this.getRevisionDetails({
-            app: application.metadata.name as string,
-            appNamespace: options.appNamespace,
-            instanceName: application.metadata.instance.name,
-            revisionID,
-          }),
-        );
-      }
-
-      const multiSourceApp = options.apps.find(app =>
-        app?.status?.history?.find(h => h?.revisions?.includes(revisionID)),
-      );
-
-      if (multiSourceApp) {
-        const history = multiSourceApp.status?.history ?? [];
-        const relevantHistories = history.filter(h =>
-          h?.revisions?.includes(revisionID),
-        );
-
-        relevantHistories.forEach(h => {
-          const revisionSourceIndex = h.revisions?.indexOf(revisionID);
-          promises.push(
-            this.getRevisionDetails({
-              app: multiSourceApp.metadata.name as string,
-              appNamespace: options.appNamespace,
-              instanceName: multiSourceApp.metadata.instance.name,
-              revisionID,
-              sourceIndex: revisionSourceIndex,
-            }),
-          );
-        });
-      }
-    });
-    return Promise.all(promises);
-  }
-
-  async getApplication(options: GetApplicationOptions): Promise<Application> {
-    const instanceName = getInstanceNameFromUrl(options.url);
-
-    if (!DEV_INSTANCE_APPLICATIONS[instanceName]) {
-      throw new Error(
-        `Failed to fetch Application from Instance ${instanceName} : ArgoCD Instance ${instanceName} not found`,
-      );
-    }
-
-    const result = DEV_INSTANCE_APPLICATIONS[instanceName].filter(
-      app => app.metadata.name === options.appName,
-    )[0];
-    if (!result) {
-      throw new Error(
-        `Failed to fetch data, status 403: Insufficient permissions for ArgoCD server`,
-      );
-    }
-
-    return result;
-  }
-
-  async findApplications(
-    options: FindApplicationsOptions,
-  ): Promise<InstanceApplications[]> {
-    const result: InstanceApplications[] = [];
-    for (const [instanceName, apps] of Object.entries(
-      DEV_INSTANCE_APPLICATIONS,
-    )) {
-      const matchingApps = apps.filter(
-        app =>
-          app.metadata.name === options.appName &&
-          app.metadata.name !== undefined,
-      );
-
-      if (matchingApps.length !== 0) {
-        result.push({
-          name: instanceName,
-          url: matchingApps[0].metadata.instance.url,
-          appName: [options.appName],
-          applications: matchingApps,
-        });
-      }
-    }
-    return result;
-  }
-}
 
 const combinedArgocdConfig = {
   argocd: {
@@ -239,47 +107,6 @@ const argocdDevModule = createFrontendModule({
   ],
 });
 
-const mockKubernetesResources = Object.entries(mockArgoResources).flatMap(
-  ([type, resources]) => {
-    if (customResourceTypes.map(t => t.toLocaleLowerCase()).includes(type)) {
-      return { type: 'customresources', resources };
-    }
-    return { type: type.toLocaleLowerCase('en-US'), resources };
-  },
-);
-
-const mockClusterResponse = {
-  items: [
-    {
-      cluster: { name: 'mock-cluster' },
-      resources: mockKubernetesResources,
-      podMetrics: [],
-      errors: [],
-    },
-  ],
-};
-
-const mockKubernetesClient = {
-  getWorkloadsByEntity: async () => mockClusterResponse,
-  getCustomObjectsByEntity: async () => mockClusterResponse,
-  getObjectsByEntity: async () => mockClusterResponse,
-  getClusters: async () => [
-    { name: 'mock-cluster', authProvider: 'serviceAccount' },
-  ],
-  getCluster: async () => ({
-    name: 'mock-cluster',
-    authProvider: 'serviceAccount',
-  }),
-  proxy: async () =>
-    new Response(
-      JSON.stringify({
-        kind: 'Namespace',
-        apiVersion: 'v1',
-        metadata: { name: 'mock-ns' },
-      }),
-    ),
-} as unknown as KubernetesApi;
-
 const kubernetesStubPlugin = createFrontendPlugin({
   pluginId: 'kubernetes',
   extensions: [],
@@ -294,7 +121,7 @@ const kubernetesDevModule = createFrontendModule({
         defineParams({
           api: kubernetesApiRef,
           deps: {},
-          factory: () => mockKubernetesClient,
+          factory: () => new MockKubernetesClient(mockArgoResources),
         }),
     }),
   ],
@@ -314,16 +141,7 @@ const kubernetesAuthDevModule = createFrontendModule({
         defineParams({
           api: kubernetesAuthProvidersApiRef,
           deps: {},
-          factory: () => ({
-            decorateRequestBodyForAuth: async () => ({
-              entity: {
-                apiVersion: 'v1',
-                kind: 'xyz',
-                metadata: { name: 'hey' },
-              },
-            }),
-            getCredentials: async () => ({}),
-          }),
+          factory: () => mockKubernetesAuthProviderApi,
         }),
     }),
   ],
