+++ workspaces/argocd/plugins/argocd/package.json
@@ -51,6 +51,7 @@
     "@backstage/catalog-model": "^1.9.0",
     "@backstage/core-components": "^0.18.11",
     "@backstage/core-plugin-api": "^1.12.7",
+    "@backstage/frontend-plugin-api": "^0.17.1",
     "@backstage/plugin-catalog-react": "^3.1.0",
     "@backstage/plugin-kubernetes-react": "^0.5.20",
     "@backstage/plugin-permission-react": "^0.5.2",
@@ -75,6 +76,11 @@
     "@backstage/cli": "^0.36.3",
     "@backstage/config": "^1.3.8",
     "@backstage/dev-utils": "^1.1.24",
+    "@backstage/frontend-defaults": "^0.5.2",
+    "@backstage/frontend-test-utils": "^0.6.0",
+    "@backstage/plugin-catalog": "^2.0.5",
+    "@backstage/plugin-permission-common": "^0.9.9",
+    "@backstage/plugin-user-settings": "^0.9.3",
     "@backstage/test-utils": "^1.7.19",
     "@playwright/test": "1.61.1",
     "@testing-library/dom": "^10.4.1",
