+++ workspaces/argocd/plugins/argocd/report.api.md
@@ -5,26 +5,6 @@
 ```ts
 import { FrontendPlugin } from '@backstage/frontend-plugin-api';
 
-// Warning: (ae-missing-release-tag) "DeploymentLifecycleProps" is part of the package's API, but it is missing a release tag (@alpha, @beta, @public, or @internal)
-//
-// @public (undocumented)
-export interface DeploymentLifecycleProps {
-  // (undocumented)
-  showInstance?: boolean;
-  // (undocumented)
-  showServer?: boolean;
-}
-
-// Warning: (ae-missing-release-tag) "DeploymentSummaryProps" is part of the package's API, but it is missing a release tag (@alpha, @beta, @public, or @internal)
-//
-// @public (undocumented)
-export interface DeploymentSummaryProps {
-  // (undocumented)
-  showInstance?: boolean;
-  // (undocumented)
-  showServer?: boolean;
-}
-
 // @public
 const plugin: FrontendPlugin;
 export default plugin;
