+++ workspaces/argocd/plugins/argocd/src/index.ts
@@ -15,6 +15,3 @@
  */
 
 export { default } from './plugin';
-
-export type { DeploymentSummaryProps } from './components/DeploymentSummary/DeploymentSummary';
-export type { DeploymentLifecycleProps } from './components/DeploymentLifeCycle/DeploymentLifecycle';
